xkb: move it under Xext/
