+++ Xext/meson.build
@@ -51,6 +51,7 @@ libxserver_xext = static_library('xserver_xext',
 
 subdir('xfixes')
 subdir('xinput')
+subdir('xkeyboard')
 subdir('damage')
 subdir('doublebuffer')
 subdir('dri2')
