+++ Xext/xinput/exevents.c
@@ -102,7 +102,7 @@ SOFTWARE.
 #include "os/bug_priv.h"
 #include "os/log_priv.h"
 #include "os/osdep.h"
-#include "xkb/xkbsrv_priv.h"
+#include "Xext/xkeyboard/xkbsrv_priv.h"
 
 #include "inputstr.h"
 #include "windowstr.h"
