+++ Xext/xkeyboard/XKBMisc.c
@@ -31,8 +31,8 @@ THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <X11/Xproto.h>
 #include <X11/keysym.h>
 
-#include "xkb/xkbfmisc_priv.h"
-#include "xkb/xkbsrv_priv.h"
+#include "xkbfmisc_priv.h"
+#include "xkbsrv_priv.h"
 
 #include "misc.h"
 #include "inputstr.h"
