+++ Xext/xkeyboard/ddxLoad.c
@@ -41,10 +41,10 @@ THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include "dix/dix_priv.h"
 #include "os/log_priv.h"
 #include "os/osdep.h"
-#include "xkb/xkbfile_priv.h"
-#include "xkb/xkbfmisc_priv.h"
-#include "xkb/xkbrules_priv.h"
-#include "xkb/xkbsrv_priv.h"
+#include "xkbfile_priv.h"
+#include "xkbfmisc_priv.h"
+#include "xkbrules_priv.h"
+#include "xkbsrv_priv.h"
 
 #include "inputstr.h"
 #include "scrnintstr.h"
