+++ Xext/xkeyboard/meson.build
@@ -22,7 +22,7 @@ srcs_xkb = [
     'XKBMAlloc.c',
 ]
 
-libxserver_xkb = static_library('xserver_xkb',
+libxserver_xkeyboard = static_library('xserver_xkeyboard',
     srcs_xkb,
     include_directories: inc,
     dependencies: common_dep,
@@ -34,7 +34,7 @@ srcs_xkb_stubs = [
     'ddxVT.c',
 ]
 
-libxserver_xkb_stubs = static_library('xserver_xkb_stubs',
+libxserver_xkeyboard_stubs = static_library('xserver_xkeyboard_stubs',
     srcs_xkb_stubs,
     include_directories: inc,
     dependencies: common_dep,
