+++ Xext/xkeyboard/xkbUtils.c
@@ -60,7 +60,7 @@ DEALINGS IN THE SOFTWARE.
 
 #include "dix/input_priv.h"
 #include "os/log_priv.h"
-#include "xkb/xkbsrv_priv.h"
+#include "xkbsrv_priv.h"
 
 #include "os.h"
 #include "misc.h"
