+++ Xext/xkeyboard/xkbfmisc.c
@@ -36,8 +36,8 @@
 #include <X11/keysym.h>
 #include <X11/Xproto.h>
 
-#include "xkb/xkbfmisc_priv.h"
-#include "xkb/xkbout_priv.h"
+#include "xkbfmisc_priv.h"
+#include "xkbout_priv.h"
 
 #include "misc.h"
 #include "inputstr.h"
