+++ Xext/xkeyboard/xkbsrv_priv.h
@@ -9,7 +9,7 @@
 #include <X11/Xdefs.h>
 #include <X11/Xmd.h>
 
-#include "xkb/xkbrules_priv.h"
+#include "xkbrules_priv.h"
 
 #include "dix.h"
 #include "input.h"
