+++ Xext/xkeyboard/xkbtext.c
@@ -34,7 +34,7 @@
 #include <X11/Xproto.h>
 #include <X11/extensions/XKMformat.h>
 
-#include "xkb/xkbtext_priv.h"
+#include "xkbtext_priv.h"
 
 #include "misc.h"
 #include "inputstr.h"
