+++ Xext/xkeyboard/xkmread.c
@@ -35,9 +35,9 @@
 #include <X11/extensions/XKMformat.h>
 
 #include "os/log_priv.h"
-#include "xkb/xkbfile_priv.h"
-#include "xkb/xkbfmisc_priv.h"
-#include "xkb/xkbsrv_priv.h"
+#include "xkbfile_priv.h"
+#include "xkbfmisc_priv.h"
+#include "xkbsrv_priv.h"
 
 #include "misc.h"
 #include "inputstr.h"
