+++ dix/devices.c
@@ -69,7 +69,7 @@ SOFTWARE.
 #include "os/bug_priv.h"
 #include "os/log_priv.h"
 #include "os/osdep.h"
-#include "xkb/xkbsrv_priv.h"
+#include "Xext/xkeyboard/xkbsrv_priv.h"
 
 #include "misc.h"
 #include "resource.h"
