+++ hw/kdrive/ephyr/meson.build
@@ -56,7 +56,7 @@ xephyr_server = executable(
         libxserver_fb,
         libxserver,
         libxserver_config,
-        libxserver_xkb_stubs,
+        libxserver_xkeyboard_stubs,
         libxserver_xi_stubs,
         libxserver_glx,
         libglxvnd,
