+++ hw/xfree86/common/xf86Config.c
@@ -54,7 +54,7 @@
 #include "include/extinit.h"
 #include "os/log_priv.h"
 #include "os/osdep.h"
-#include "xkb/xkbsrv_priv.h"
+#include "Xext/xkeyboard/xkbsrv_priv.h"
 #ifdef DPMSExtension
 #include "Xext/dpms/dpms_priv.h"
 #endif
