+++ hw/xfree86/xkb/xkbVT.c
@@ -32,7 +32,7 @@ THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <X11/extensions/XI.h>
 
 #include "hw/xfree86/common/action_priv.h"
-#include "xkb/xkbsrv_priv.h"
+#include "Xext/xkeyboard/xkbsrv_priv.h"
 
 #include "inputstr.h"
 #include "scrnintstr.h"
