+++ hw/xnest/meson.build
@@ -43,7 +43,7 @@ executable(
         libxserver_main,
         libxserver,
         libxserver_xi_stubs,
-        libxserver_xkb_stubs,
+        libxserver_xkeyboard_stubs,
     ],
     c_args: [ '-DHAVE_XNEST_CONFIG_H', '-DDISABLE_EXT_COMPOSITE', '-DISABLE_EXT_MITSHM' ],
     install: true,
