+++ hw/xwin/InitOutput.c
@@ -38,7 +38,7 @@ from The Open Group.
 #include "os/ddx_priv.h"
 #include "os/log_priv.h"
 #include "os/osdep.h"
-#include "xkb/xkbsrv_priv.h"
+#include "Xext/xkeyboard/xkbsrv_priv.h"
 #ifdef DPMSExtension
 #include "Xext/dpms/dpms_priv.h"
 #endif
